+++ config/runtime.exs
@@ -153,19 +153,31 @@ end
 
 # Can be generated  with `Base.encode64(:crypto.strong_rand_bytes(32))` from
 # iex shell or `openssl rand -base64 32` from command line.
-totp_vault_key = get_var_from_path_or_env(config_dir, "TOTP_VAULT_KEY", nil)
+totp_vault_key =
+  if totp_vault_key_base64 = get_var_from_path_or_env(config_dir, "TOTP_VAULT_KEY") do
+    case Base.decode64(totp_vault_key_base64) do
+      {:ok, totp_vault_key} ->
+        if byte_size(totp_vault_key) == 32 do
+          totp_vault_key
+        else
+          raise ArgumentError, """
+          TOTP_VAULT_KEY must be Base64 encoded 32 bytes, e.g. `openssl rand -base64 32`.
+          Got Base64 encoded #{byte_size(totp_vault_key)} bytes.
+          More info: https://github.com/plausible/community-edition/tree/v2.1.1#quick-start
+          """
+        end
 
-case totp_vault_key do
-  nil ->
-    raise "TOTP_VAULT_KEY configuration option is required. See https://github.com/plausible/community-edition/tree/v2.1.0?tab=readme-ov-file#quick-start"
-
-  key ->
-    if byte_size(Base.decode64!(key)) != 32 do
-      raise "TOTP_VAULT_KEY must exactly 32 bytes long. See https://github.com/plausible/community-edition/tree/v2.1.0?tab=readme-ov-file#quick-start"
+      :error ->
+        raise ArgumentError, """
+        TOTP_VAULT_KEY must be Base64 encoded 32 bytes, e.g. `openssl rand -base64 32`
+        More info: https://github.com/plausible/community-edition/tree/v2.1.1#quick-start
+        """
     end
-end
+  else
+    Plug.Crypto.KeyGenerator.generate(secret_key_base, "totp", length: 32, iterations: 100_000)
+  end
 
-### Mandatory params End
+config :plausible, Plausible.Auth.TOTP, vault_key: totp_vault_key
 
 build_metadata_raw = get_var_from_path_or_env(config_dir, "BUILD_METADATA", "{}")
 
@@ -196,8 +208,6 @@ runtime_metadata = [
 ]
 
 config :plausible, :runtime_metadata, runtime_metadata
-
-config :plausible, Plausible.Auth.TOTP, vault_key: totp_vault_key
 
 sentry_dsn = get_var_from_path_or_env(config_dir, "SENTRY_DSN")
 honeycomb_api_key = get_var_from_path_or_env(config_dir, "HONEYCOMB_API_KEY")
@@ -504,13 +514,24 @@ case mailer_adapter do
 
     if relay = get_var_from_path_or_env(config_dir, "SMTP_HOST_ADDR") do
       port = get_int_from_path_or_env(config_dir, "SMTP_HOST_PORT", 25)
-      username = get_var_from_path_or_env(config_dir, "SMTP_USER_NAME")
-      password = get_var_from_path_or_env(config_dir, "SMTP_USER_PWD")
+      config :plausible, Plausible.Mailer, relay: relay, port: port
+    end
 
-      config :plausible, Plausible.Mailer,
-        auth: [username: username, password: password],
-        relay: relay,
-        port: port
+    username = get_var_from_path_or_env(config_dir, "SMTP_USER_NAME")
+    password = get_var_from_path_or_env(config_dir, "SMTP_USER_PWD")
+
+    cond do
+      username && password ->
+        config :plausible, Plausible.Mailer, auth: [username: username, password: password]
+
+      username || password ->
+        raise ArgumentError, """
+        Both SMTP_USER_NAME and SMTP_USER_PWD must be set for SMTP authentication.
+        Please provide values for both environment variables.
+        """
+
+      _both_nil = true ->
+        nil
     end
 
   "Bamboo.LocalAdapter" ->